api/requirements.txt
# HY-Motion Retargeting API Requirements
# These are already installed in the Docker image, listed for reference

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.32.0

# ML/CUDA (already in base image)
# torch>=2.0.0
# torchvision

# HY-Motion dependencies
huggingface_hub==0.30.0
torchdiffeq==0.2.5
accelerate==0.30.1
diffusers==0.26.3
transformers==4.53.3
einops==0.8.1
safetensors==0.5.3
scipy>=1.10.0
transforms3d==0.4.2
PyYAML==6.0
omegaconf==2.3.0
click==8.1.3

# FBX SDK
fbxsdkpy

# Utilities
numpy>=1.24.0,<2.0
